nowait_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.9.1
python-jose[cryptography]==3.3.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
httpx==0.27.2
python-dotenv==1.0.1
